+++ lib/presentation/screens/catalog/widgets/inputs_user_v2/input_text_v2.dart
@@ -1,7 +1,5 @@
 import 'package:finniu/presentation/providers/settings_provider.dart';
-import 'package:finniu/presentation/screens/catalog/widgets/text_poppins.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/svg.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class InputTextFileUserProfile extends ConsumerWidget {
@@ -75,108 +73,6 @@ class InputTextFileUserProfile extends ConsumerWidget {
             ),
             focusedErrorBorder: const OutlineInputBorder(
               borderRadius: BorderRadius.all(Radius.circular(25.0)),
-              borderSide: BorderSide.none,
-            ),
-          ),
-          validator: validator,
-          keyboardType: isNumeric ? TextInputType.number : TextInputType.text,
-        ),
-        const SizedBox(height: 15),
-      ],
-    );
-  }
-}
-
-class InputTextNetworkUserProfile extends ConsumerWidget {
-  const InputTextNetworkUserProfile({
-    super.key,
-    required this.controller,
-    required this.hintText,
-    required this.validator,
-    required this.urlIcon,
-    required this.textTitle,
-    this.isNumeric = false,
-  });
-
-  final TextEditingController controller;
-  final String? Function(String?)? validator;
-  final String hintText;
-  final bool isNumeric;
-  final String urlIcon;
-  final String textTitle;
-
-  final int hintDark = 0xFF989898;
-  final int hintLight = 0xFF989898;
-  final int fillDark = 0xFF222222;
-  final int fillLight = 0xFFF7F7F7;
-  final int textDark = 0xFFFFFFFF;
-  final int textLight = 0xFF000000;
-  final int borderColorDark = 0xFFA2E6FA;
-  final int borderColorLight = 0xFF0D3A5C;
-  final int iconColorDark = 0xFFA2E6FA;
-  final int iconColorLight = 0xFF0D3A5C;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final isDarkMode = ref.watch(settingsNotifierProvider).isDarkMode;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            SvgPicture.asset(
-              urlIcon,
-              width: 27,
-              height: 27,
-              color: isDarkMode ? Color(iconColorDark) : Color(iconColorLight),
-            ),
-            const SizedBox(width: 10),
-            TextPoppins(text: textTitle, fontSize: 12),
-          ],
-        ),
-        const SizedBox(height: 10),
-        TextFormField(
-          controller: controller,
-          style: TextStyle(
-            fontSize: 12,
-            color: isDarkMode ? Color(textDark) : Color(textLight),
-            fontWeight: FontWeight.w400,
-            fontFamily: "Poppins",
-          ),
-          decoration: InputDecoration(
-            hintStyle: TextStyle(
-              fontSize: 12,
-              color: isDarkMode ? Color(hintDark) : Color(hintLight),
-              fontWeight: FontWeight.w400,
-              fontFamily: "Poppins",
-            ),
-            hintText: hintText,
-            fillColor: isDarkMode ? Color(fillDark) : Color(fillLight),
-            filled: true,
-            border: const OutlineInputBorder(
-              borderRadius: BorderRadius.all(Radius.circular(5.0)),
-              borderSide: BorderSide.none,
-            ),
-            enabledBorder: const OutlineInputBorder(
-              borderRadius: BorderRadius.all(Radius.circular(5.0)),
-              borderSide: BorderSide.none,
-            ),
-            focusedBorder: OutlineInputBorder(
-              borderRadius: const BorderRadius.all(Radius.circular(5.0)),
-              borderSide: BorderSide(
-                color: isDarkMode
-                    ? Color(borderColorDark)
-                    : Color(borderColorLight),
-                width: 1,
-              ),
-            ),
-            errorBorder: const OutlineInputBorder(
-              borderRadius: BorderRadius.all(Radius.circular(5.0)),
-              borderSide: BorderSide.none,
-            ),
-            focusedErrorBorder: const OutlineInputBorder(
-              borderRadius: BorderRadius.all(Radius.circular(5.0)),
               borderSide: BorderSide.none,
             ),
           ),
